Add more profiler markers

+++ src/xrCore/xrstring.h
@@ -155,7 +155,7 @@ public:
 
 	template<>
 	static void FixSlashes<xr_string>(xr_string& InStr) {
-		for (int i = 0; i < InStr.size(); ++i)
+		for (size_t i = 0; i < InStr.size(); ++i)
 		{
 			if (InStr[i] == '/')
 			{
